Restructure files, hopefully nothing breaks >.<
