+++ .travis.yml
@@ -3,7 +3,7 @@ python:
   - "3.6"
 install:
   - make prepare
-  - make develop extras=[aws,google] # adding extras to avoid import errors
+  - make develop extras=[aws,google,wdl] # adding extras to avoid import errors
 script:
   - TOIL_TEST_QUICK=True make test_offline
 env:
