+++ src/toil/test/wdl/conftest.py
@@ -17,7 +17,7 @@
 collect_ignore = []
 
 try:
-    import toil.wdl.wdl_parser
-    print(toil.wdl.wdl_parser.__file__)  # keep this import from being removed
+    import wdlparse
+    print(wdlparse.__file__)  # keep this import from being removed
 except ImportError:
     collect_ignore.append("toilwdl.py")
